requirements.txt
streamlit==1.56.0
openai==2.30.0
pypdf
pdfplumber
pytesseract
pdf2image
Pillow
fpdf2
pytest==9.0.2
tqdm==4.67.1
rich>=13.7.0
langdetect==1.0.9
python-dotenv>=1.0.0

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.12.0

# Deadline & Notification System
twilio>=8.10.0
sendgrid>=6.10.0
apscheduler>=3.10.4
pytz>=2024.1
structlog>=24.1.0

# Authentication
pyjwt>=2.8.0
bcrypt>=4.1.0

# Testing
pytest-mock>=3.12.0
responses>=0.24.0
pytest-cov>=4.1.0

# Analytics & Visualization
pandas>=2.0.0
plotly>=5.14.0
numpy>=1.24.0
scikit-learn>=1.3.0

# RAG & Interactive Chat
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2

# Audio & Accessibility
gTTS>=2.5.1

# ==================== Observability & Monitoring ====================
# Prometheus Metrics
prometheus-client>=0.17.0

# Structured Logging
structlog>=24.1.0
python-json-logger>=2.0.0

# Distributed Tracing
jaeger-client>=4.8.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-prometheus>=0.41b0
opentelemetry-exporter-jaeger-thrift>=1.20.0
opentelemetry-instrumentation-flask>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-sqlalchemy>=0.41b0
opentelemetry-instrumentation-redis>=0.41b0

# Elasticsearch & Log Aggregation
elasticsearch>=8.0.0

# Production Server (for Prometheus metrics endpoint)
gunicorn>=21.0.0
Flask>=3.0.0

# Code Quality & Security
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
bandit>=1.7.0

# Performance Profiling
memory-profiler>=0.61.0
py-spy>=0.3.14

# HTTP client with retries
requests>=2.31.0
tenacity>=8.2.0

# Configuration Management
python-dotenv>=1.0.0
pydantic>=2.0.0

# Kubernetes Python Client (for operators/controllers)
kubernetes>=28.0.0
